requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
schedule>=1.2.0
pandas>=2.1.0
unidecode>=1.3.0
Pillow>=10.0.0
python-dateutil>=2.8.0
pytz>=2023.3
tenacity>=8.2.0
